# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Fantasm API
  version: 2.0.0
  description: |
    JSON API of the Fantasm backend. Every route lives under `/api`.

    **Authentication.** Sign-in is OIDC only (Google, Microsoft Entra). A successful callback sets an
    opaque, HttpOnly, SameSite=Lax session cookie (`__Host-fantasm_session` over HTTPS). There are no passwords.

    **CSRF.** Unsafe methods (POST, PUT, PATCH, DELETE) must come from the site's own origin: a request with a
    foreign `Origin` or `Sec-Fetch-Site: cross-site` is answered with 403.

    **Errors.** Every failure has the same body, see `Error`, including 404 for an unknown path and 405 (with `Allow`) for a wrong method. Hidden, deleted and nonexistent content all
    answer 404 identically, so existence is never revealed. Unexpected failures are a generic 500 with a `requestId`.

    **Pagination.** Lists are keyset-paginated: pass the previous page's `nextCursor` as `cursor`. Cursors are
    signed and bound to the listing they came from. `limit` is clamped to the documented maximum.

    **Limits.** Request bodies are capped (1 MiB), unknown JSON fields are rejected (400), and every route is
    rate limited per user (signed in) or per address (anonymous). Limited answers carry `Retry-After`.
servers:
  - url: /
security: []
...
paths:
...
  /api/auth/{provider}/login:
    get:
      tags: [auth]
      summary: Begin the OIDC authorization-code flow (PKCE, state, nonce)
      description: Stores a short-lived, hashed login attempt and redirects to the provider. Rate limited per address.
      parameters:
        - $ref: '#/components/parameters/Provider'
      responses:
        '302': { description: "Redirect to the provider's authorization screen." }
        '303': { description: "Browser navigation only. Redirect to `/login?error=provider_unavailable`." }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/{provider}/callback:
    get:
      tags: [auth]
      summary: Complete sign-in
      description: |
        Verifies state, nonce, PKCE and the ID token, then creates a session and sets the cookie.
        A previous session cookie is revoked (session rotation).
      parameters:
        - $ref: '#/components/parameters/Provider'
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: code, in: query, required: true, schema: { type: string } }
      responses:
        '303': { description: "Redirect to `/` on success, or to `/login?error=...` on failure." }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Sign out of this device
      responses:
        '204': { description: "Session revoked and cookie cleared." }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/logout-all:
    post:
      tags: [auth]
      summary: Sign out everywhere
      security: [{ session: [] }]
      responses:
        '204': { description: "All of the user's sessions revoked." }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me:
    get:
      tags: [me]
      summary: Current profile
      description: Anonymous viewers get 401, which the frontend treats as "signed out".
      security: [{ session: [] }]
      responses:
        '200':
          description: Profile and organization memberships.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [me]
      summary: Edit own profile
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                handle: { type: string, pattern: '^[a-z0-9_][a-z0-9_-]{2,39}$' }
                name: { type: string }
                bio: { type: string }
                faculty: { type: string }
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Validation' }
  /api/me/sessions:
    get:
      tags: [me]
      summary: Active sessions of the user
      security: [{ session: [] }]
      responses:
        '200':
          description: Sessions, the current one marked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me/sessions/{id}:
    delete:
      tags: [me]
      summary: Revoke one of the user's sessions
      security: [{ session: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: "Revoked." }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/users/{handle}/role:
    patch:
      tags: [admin]
      summary: Change a user's role
      description: Admins only. The target's sessions are revoked so the new role takes effect at once.
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/Handle'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { type: string, enum: [USER, MODERATOR, ADMIN] }
      responses:
        '200':
          description: Role changed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  handle: { type: string }
                  role: { type: string, enum: [USER, MODERATOR, ADMIN] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-fantasm_session
      description: Opaque session token set by the OIDC callback. Only its SHA-256 hash is stored.
  parameters:
    Provider:
      name: provider
      in: path
      required: true
      schema: { type: string, enum: [google, entra] }
...
    Handle:
      name: handle
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9_][a-z0-9_-]{2,39}$' }
...
  responses:
    Unauthorized:
      description: Not signed in, or the session expired.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Signed in but not allowed, or a cross-origin request.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such resource, or one the viewer may not know about.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Validation:
      description: The input is well-formed but not acceptable; `fields` names what to fix.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait. }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [invalid_request, unauthorized, forbidden, not_found, conflict, rate_limited, payload_too_large, validation_failed, unsupported_media_type, method_not_allowed, not_implemented, internal_error]
        fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
...
    Membership:
      type: object
      properties:
        organizationId: { type: string, example: naukma }
        name: { type: string }
        badge: { type: string }
        capabilities: { type: array, items: { type: string } }
        benefits:
          type: object
          properties:
            voteWeight: { type: integer }
            rankingMultiplier: { type: number }
            karmaMultiplier: { type: number }
            skipPremoderation: { type: boolean }
    User:
      type: object
      required: [id, handle, name, email, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        handle: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: string }
        bio: { type: string }
        faculty: { type: string }
        role: { type: string, enum: [USER, MODERATOR, ADMIN] }
        karma: { type: integer }
        verified: { type: boolean }
        memberships: { type: array, items: { $ref: '#/components/schemas/Membership' } }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      properties:
        id: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        userAgent: { type: string }
        current: { type: boolean }
...